Refactor all elasticsearch controllers to use one client
